Replace the RolePahResolver with the enhanced ContextPathResolver ...
This allows for multiple references to be inject at the very initial steps of
the pairing and all the references are placed into the node_contect of each
peer. If the reference is local, it's replaced with the real object, if it's
remote, it will be replaced with a Proxy, acting as a facade.